Split playback preferences into normal and advanced screens
