+++ app/src/main/res/values/strings.xml
@@ -506,6 +506,7 @@
     <string name="clear_image_cache_content">Used: %1$s</string>
     <string name="random">Random</string>
     <string name="unreleased">Not yet released</string>
+    <string name="pref_playback_advanced">Advanced playback preferences</string>
     <plurals name="seconds">
         <item quantity="one">%1$s second</item>
         <item quantity="other">%1$s seconds</item>
